menambahkan pesan eror id sama di post dan id kosong di put

+++ src/main/java/pws/a/ppp/ProductServiceController.java
@@ -44,18 +44,32 @@ public class ProductServiceController {
    
    @RequestMapping(value = "/products/{id}", method = RequestMethod.PUT)
    public ResponseEntity<Object> updateProduct(@PathVariable("id") String id, @RequestBody Product product) 
-   { 
-      productRepo.remove(id);
-      product.setId(id);
-      productRepo.put(id, product);
-      return new ResponseEntity<>("Product is updated successsfully", HttpStatus.OK);
+   {
+      if(!productRepo.containsKey(id))
+      { 
+        return new ResponseEntity<>("Product not found", HttpStatus.NOT_FOUND);
+      }
+      else
+      {
+        productRepo.remove(id);
+        product.setId(id);
+        productRepo.put(id, product);
+        return new ResponseEntity<>("Product is updated successsfully", HttpStatus.OK);
+      }
    }
    
    @RequestMapping(value = "/products", method = RequestMethod.POST)
    public ResponseEntity<Object> createProduct(@RequestBody Product product) 
    {
-      productRepo.put(product.getId(), product);
-      return new ResponseEntity<>("Product is created successfully", HttpStatus.CREATED);
+      if (productRepo.containsKey(product.getId()))
+      {
+        return new ResponseEntity<>("Cannot add same id", HttpStatus.OK);
+      }
+      else
+      {
+        productRepo.put(product.getId(), product);      
+        return new ResponseEntity<>("Product is created successfully", HttpStatus.CREATED);
+      }
    }
    
    @RequestMapping(value = "/products")
